Initial check-in
Rough initial version of programmer. Expectation was that fx_uploader would send data to a known location, and this would access that... however, liberis clears out all of memory, so that doesn't work.
A hacked binary of this version does work however; it does need some way of loading the binary payload data though.